feat: expose GenerateNoProxy func
