+++ pkg/handlers/generic/mutation/httpproxy/systemd_proxy_config_test.go
@@ -119,7 +119,8 @@ Environment="no_proxy=localhost,127.0.0.1,no-proxy.example.com,no-proxy-1.exampl
 				}}
 			}
 
-			g.Expect(generateSystemdFiles(tt.vars, tt.noProxy)).Should(Equal(expectedFiles))
+			g.Expect(generateSystemdFiles(tt.vars, append(tt.noProxy, tt.vars.AdditionalNo...))).
+				Should(Equal(expectedFiles))
 		})
 	}
 }
